refactor(window_resize): remove
