+++ client/src/app/event_listeners.js
@@ -6,13 +6,10 @@ import {
 import { onStorageChange } from '../users/authentication'
 import { onWindowBeforeUnload, onNoConnection } from '../util/fetch_nonblocking'
 import { onGlobalKeyDown } from './keyboard_commands'
-import { onResize } from './window_resize'
 import { addPageVisibilityChangeListeners } from './focus'
 
 export function addEventListeners () {
   window.addEventListener('storage', onStorageChange)
-
-  window.addEventListener('resize', onResize)
 
   window.addEventListener('pointerdown', onBodyMouseDown)
   window.addEventListener('pointermove', onBodyMouseMove)
