+++ client/src/info_bubble/InfoBubbleLower/DescriptionPanel.tsx
@@ -9,7 +9,7 @@ import Transition, {
   type TransitionStatus
 } from 'react-transition-group/Transition'
 
-import { getStreetSectionTop } from '~/src/app/window_resize'
+import { app } from '~/src/preinit/app_settings'
 import Triangle from './Triangle'
 import './DescriptionPanel.scss'
 
@@ -31,6 +31,34 @@ Record<TransitionStatus, React.CSSProperties>
     opacity: 1,
     transform: 'none'
   }
+}
+
+// TODO: less reliance on querying other DOM elements, if possible
+// TODO: document all magic numbers
+export function getStreetSectionTop (): number {
+  let streetSectionTop
+
+  // not _always_ equal to 100vh, but for our purposes, it is
+  const viewportHeight = window.innerHeight
+
+  // TODO this will always be the same, since in CSS this is a fixed number.
+  const streetSectionHeight = document.querySelector(
+    '#street-section-inner'
+  )?.offsetHeight
+
+  // TODO const
+  if (viewportHeight - streetSectionHeight > 450) {
+    streetSectionTop =
+      (viewportHeight - streetSectionHeight - 450) / 2 + 450 + 80
+  } else {
+    streetSectionTop = viewportHeight - streetSectionHeight + 70
+  }
+
+  if (app.readOnly) {
+    streetSectionTop += 80
+  }
+
+  return streetSectionTop
 }
 
 interface DescriptionPanelProps {
